# offerpilot-openapi.yaml
openapi: 3.0.3
info:
  title: OfferPilot API
  version: 1.0.0
  description: OfferPilot 前端实习求职训练工作台后端接口。可直接导入 Apifox。
servers:
  - url: /api/v1
    description: API 前缀
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [账号与用户]
      summary: 用户注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  example: 前端实习生
                email:
                  type: string
                  format: email
                  example: name@example.com
                password:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: 邮箱已注册或参数错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [账号与用户]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: name@example.com
                password:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: 登录失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me:
    get:
      tags: [账号与用户]
      summary: 获取当前用户
      responses:
        "200":
          description: 当前用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [账号与用户]
      summary: 更新个人设置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 前端冲刺选手
                goal:
                  type: string
                  example: 每天复习 5 道题，并完成 1 次投递跟进
                notifications:
                  type: boolean
                  example: true
      responses:
        "200":
          description: 更新后的用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
...
  /questions:
    get:
      tags: [面试题库]
      summary: 获取题目列表
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
          description: 搜索标题、答案、标签
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/QuestionCategory"
        - name: difficulty
          in: query
          schema:
            $ref: "#/components/schemas/Difficulty"
        - name: starred
          in: query
          schema:
            type: boolean
        - name: mastered
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 题目分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: "#/components/schemas/Question"
                          pagination:
                            $ref: "#/components/schemas/Pagination"
    post:
      tags: [面试题库]
      summary: 新增题目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionCreate"
      responses:
        "200":
          description: 新增后的题目
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Question"
...
  /algorithms:
    get:
      tags: [算法练习]
      summary: 获取算法题列表
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProblemStatus"
        - name: difficulty
          in: query
          schema:
            $ref: "#/components/schemas/Difficulty"
        - name: keyword
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 算法题分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: "#/components/schemas/AlgorithmProblem"
                          pagination:
                            $ref: "#/components/schemas/Pagination"
    post:
      tags: [算法练习]
      summary: 新增算法题
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlgorithmProblemCreate"
      responses:
        "200":
          description: 新增后的算法题
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AlgorithmProblem"
...
  /experiences:
    get:
      tags: [面经笔记]
      summary: 获取面经列表
      parameters:
        - name: round
          in: query
          schema:
            type: string
            enum: [一面, 二面, HR 面, Offer]
        - name: result
          in: query
          schema:
            $ref: "#/components/schemas/ExperienceResult"
        - name: keyword
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 面经分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: "#/components/schemas/Experience"
                          pagination:
                            $ref: "#/components/schemas/Pagination"
    post:
      tags: [面经笔记]
      summary: 新增面经
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceCreate"
      responses:
        "200":
          description: 新增后的面经
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Experience"
...
  /experiences/import:
    post:
      tags: [面经笔记]
      summary: 导入面经 JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/ExperienceCreate"
      responses:
        "200":
          description: 导入结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          created:
                            type: integer
                            example: 3
                          updated:
                            type: integer
                            example: 0
                          failed:
                            type: integer
                            example: 0
...
  /applications:
    get:
      tags: [投递看板]
      summary: 获取投递列表
      parameters:
        - name: stage
          in: query
          schema:
            $ref: "#/components/schemas/ApplicationStage"
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 投递列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Application"
    post:
      tags: [投递看板]
      summary: 新增投递
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationCreate"
      responses:
        "200":
          description: 新增后的投递
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Application"
...
  /mock-interview/questions/random:
    get:
      tags: [模拟面试]
      summary: 按模式随机抽题
      parameters:
        - name: mode
          in: query
          required: true
          schema:
            type: string
            example: Vue 专项
      responses:
        "200":
          description: 随机题目
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MockQuestion"
  /mock-interview/sessions:
    get:
      tags: [模拟面试]
      summary: 获取模拟面试历史
      parameters:
        - name: mode
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MockSessionStatus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 会话历史
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: "#/components/schemas/MockSession"
                          pagination:
                            $ref: "#/components/schemas/Pagination"
    post:
      tags: [模拟面试]
      summary: 创建模拟面试会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode, questionIds]
              properties:
                mode:
                  type: string
                  example: Vue 专项
                questionIds:
                  type: array
                  items:
                    type: string
                  example: [mock_q_001]
      responses:
        "200":
          description: 新建会话
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MockSession"
...
  /notifications:
    get:
      tags: [通知提醒]
      summary: 获取通知列表
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 通知分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: "#/components/schemas/Notification"
                          pagination:
                            $ref: "#/components/schemas/Pagination"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: 页码
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
      description: 每页数量
...
  responses:
    Unauthorized:
      description: 未登录或 token 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 40100
        message:
          type: string
          example: 未登录或 token 无效
        data:
          nullable: true
          example: null
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
    AuthResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: "#/components/schemas/User"
                accessToken:
                  type: string
                  example: jwt-token
    User:
      type: object
      properties:
        id:
          type: string
          example: user_001
        name:
          type: string
          example: 前端实习生
        email:
          type: string
          example: name@example.com
        goal:
          type: string
          example: 每天完成 2 小时学习与 3 道算法题
        notifications:
          type: boolean
          example: true
        createdAt:
          type: string
          example: "2026-05-14T09:30:00+08:00"
        updatedAt:
          type: string
          example: "2026-05-14T09:30:00+08:00"
    QuestionCategory:
      type: string
      enum: [CSS, JavaScript, Vue, Network, Algorithm]
      example: JavaScript
    Difficulty:
      type: string
      enum: [easy, medium, hard]
      example: medium
...
    Question:
      allOf:
        - $ref: "#/components/schemas/QuestionUpdate"
        - type: object
          properties:
            id:
              type: string
              example: q_js_001
            title:
              type: string
              example: 介绍一下事件循环
            category:
              $ref: "#/components/schemas/QuestionCategory"
            difficulty:
              $ref: "#/components/schemas/Difficulty"
            tags:
              type: array
              items:
                type: string
              example: [event-loop, promise, async]
            answer:
              type: string
              example: 同步任务先进入调用栈，异步回调分别进入宏任务或微任务队列。
            followUps:
              type: array
              items:
                type: string
              example: [Promise.then 属于什么任务？]
            pitfalls:
              type: array
              items:
                type: string
              example: [不要把 async 函数本身说成微任务。]
            mastered:
              type: boolean
              example: false
            starred:
              type: boolean
              example: true
            lastReviewedAt:
              type: string
              example: "2026-05-13"
            note:
              type: string
              example: 结合页面渲染时机再复盘一次。
            createdAt:
              type: string
              example: "2026-05-10T10:00:00+08:00"
            updatedAt:
              type: string
              example: "2026-05-14T09:30:00+08:00"
...
    ProblemStatus:
      type: string
      enum: [todo, doing, done]
      example: doing
...
    AlgorithmProblem:
      allOf:
        - $ref: "#/components/schemas/AlgorithmProblemUpdate"
        - type: object
          properties:
            id:
              type: string
              example: lc_003
            updatedAt:
              type: string
              example: "2026-05-14T09:30:00+08:00"
...
    ExperienceResult:
      type: string
      enum: [pending, passed, failed]
      example: pending
...
    ApplicationStage:
      type: string
      enum: [todo, applied, exam, interview_1, interview_2, hr, offer, closed]
      example: interview_1
...